refactor: home page component분리
