+++ src/components/common/Navigation/index.tsx
@@ -69,4 +69,4 @@ function Navigation() {
     </>
   );
 }
-export default Navigation;
+export default React.memo(Navigation);
